Add a script to install mongoose (3.1) from source
Mongoose is a lightweight web server capable of serving cgi scripts.
To provide a working instaweb function 'out of the box' for msysgit
we need to be sure there is a web server we can host the gitweb.cgi
script on.

mongoose has no dependencies, compiles successfully without patching
and results in a small (52Kb) executable so would appear to be an ideal
candidate for shipping to re-enable this missing instaweb functionality
in the deliverables.

Signed-off-by: Ciaran Jessup <ciaranj@gmail.com>